# lc-ssi-wc/openapi/swift-data-service.v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Baseline SWIFT Data Service",
    "version": "1.1.0-prototype",
    "description": "SR2026 prototype contract. BIC syntax follows ISO 9362; currency codes are supplied by the ISO 4217 Currency service. Licensed SwiftRef/RMA/SSI data remains a production prerequisite."
  },
  "servers": [
    {
      "url": "http://localhost:3100/api"
    }
  ],
...
  "paths": {
...
    "/reference/fin-resolution-catalogue": {
      "get": {
        "operationId": "listFinResolutionCatalogue",
        "description": "Returns the SR-governed outward MT3/MT4/MT7 catalogue with an explicit message-level SSI resolution scope marker and SSI-resolvable tag options.",
        "parameters": [
          {
            "name": "standardsRelease",
            "in": "query",
            "schema": { "type": "string", "default": "SR2026" }
          }
        ],
        "responses": {
          "200": {
            "description": "Parameter-driven FIN Resolution catalogue; each message is marked SSI_RESOLVABLE or NO_SSI_RESOLVABLE_FIELDS."
          }
        }
      }
    },
    "/reference/fin-tag-resolutions": {
      "post": {
        "operationId": "resolveReferenceFin5xFields",
        "description": "Primary SR2026-governed, reference-only FIN field resolution for Treasury MT3 and Trade Finance MT4/MT7. No payment execution or confirmation semantics.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "service",
                  "resolutionMode",
                  "standardsRelease",
                  "messageType",
                  "direction",
                  "businessFunction",
                  "transactionReference",
                  "currency",
                  "receiverBic",
                  "valueDate",
                  "roles"
                ],
                "properties": {
                  "service": {
                    "const": "FIN"
                  },
                  "resolutionMode": {
                    "enum": ["TREASURY", "TRADE_FINANCE"]
                  },
                  "standardsRelease": {
                    "const": "SR2026"
                  },
                  "messageType": {
                    "type": "string",
                    "pattern": "^MT[0-9]{3}$"
                  },
                  "direction": {
                    "const": "OUTGOING"
                  },
                  "businessFunction": {
                    "type": "string"
                  },
                  "sequence": {
                    "type": "string"
                  },
                  "settlementLeg": {
                    "type": "string"
                  },
                  "fieldOptions": {
                    "type": "object",
                    "additionalProperties": {
                      "type": "string"
                    }
                  },
                  "transactionReference": {
                    "type": "string"
                  },
                  "currency": {
                    "type": "string",
                    "pattern": "^[A-Z]{3}$"
                  },
                  "receiverBic": {
                    "type": "string"
                  },
                  "bookingEntity": {
                    "type": "string"
                  },
                  "valueDate": {
                    "type": "string",
                    "format": "date"
                  },
                  "roles": {
                    "type": "object",
                    "additionalProperties": {
                      "type": "string"
                    }
                  },
                  "roleEvidence": {
                    "type": "object",
                    "description": "Evidence and canonical settlement-route node identity for each role. Equal BICs are normalized only when canonicalRouteNodeId and ownerSide also match.",
                    "additionalProperties": {
                      "type": "object",
                      "properties": {
                        "ownerSide": {
                          "enum": [
                            "SENDER_SIDE",
                            "RECEIVER_SIDE",
                            "TRANSACTION_PARTY"
                          ]
                        },
                        "sourceType": { "type": "string" },
                        "sourceRecordId": { "type": "string" },
                        "version": { "type": "string" },
                        "status": { "type": "string" },
                        "approvalStatus": { "type": "string" },
                        "effectiveFrom": { "type": "string", "format": "date" },
                        "effectiveTo": { "type": "string", "format": "date" },
                        "canonicalRouteNodeId": {
                          "type": "string",
                          "description": "Identity derived from normalized BIC11, branch, account owner/servicer/identifier, currency, clearing context and instruction provenance; a BIC alone is insufficient."
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Reference-only resolvedFields result; paymentExecutable is always false.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "usage",
                    "paymentExecutable",
                    "confirmationSupported",
                    "resolvedFields"
                  ],
                  "properties": {
                    "usage": {
                      "const": "REFERENCE_ONLY"
                    },
                    "paymentExecutable": {
                      "const": false
                    },
                    "confirmationSupported": {
                      "const": false
                    },
                    "resolvedFields": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "standardsRelease",
                          "messageType",
                          "sequence",
                          "settlementLeg",
                          "tag",
                          "option",
                          "officialFieldName",
                          "officialRole",
                          "businessFunction",
                          "scopeStatus",
                          "resolutionStatus",
                          "reasonCode",
                          "resolvedValue",
                          "nvrRefs",
                          "provenance"
                        ],
                        "properties": {
                          "standardsRelease": {
                            "type": "string"
                          },
                          "messageType": {
                            "type": "string"
                          },
                          "sequence": {
                            "type": "string"
                          },
                          "settlementLeg": {
                            "type": "string"
                          },
                          "tag": {
                            "type": "string"
                          },
                          "option": {
                            "type": "string"
                          },
                          "officialFieldName": {
                            "type": "string"
                          },
                          "officialRole": {
                            "type": "string"
                          },
                          "businessFunction": {
                            "type": "string"
                          },
                          "scopeStatus": {
                            "enum": ["SSI_SUPPORTED", "OUT_OF_SSI_SCOPE"]
                          },
                          "resolutionStatus": {
                            "description": "Runtime disposition for every audited field slot. Profile-excluded slots are returned explicitly as N_A rather than omitted.",
                            "enum": [
                              "RESOLVED",
                              "NOT_REQUIRED",
                              "NO_ELIGIBLE_SSI",
                              "N_A"
                            ]
                          },
                          "reasonCode": {
                            "description": "Controlled reason such as ROUTE_COMPLETE or MESSAGE_PROFILE_EXCLUDED. A ROUTE_COMPLETE decision is attributed to ROUTE_RESOLVER/CANONICAL_ROUTE in provenance.",
                            "type": "string"
                          },
                          "resolvedValue": {
                            "type": ["string", "null"]
                          },
                          "nvrRefs": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "provenance": {
                            "description": "For NOT_REQUIRED route decisions, source is ROUTE_RESOLVER and ownerSide is CANONICAL_ROUTE; original candidate source/owner are retained separately when applicable.",
                            "type": "object"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Context, exact profile, option, or eligible SSI evidence is invalid/missing."
          }
        }
      }
    },
...
  },
...
}